fix: remove non default ipld formats in the browser
