+++ package.json
@@ -16,6 +16,7 @@
     "./src/core/runtime/libp2p-nodejs.js": "./src/core/runtime/libp2p-browser.js",
     "./src/core/runtime/preload-nodejs.js": "./src/core/runtime/preload-browser.js",
     "./src/core/runtime/repo-nodejs.js": "./src/core/runtime/repo-browser.js",
+    "./src/core/runtime/ipld-nodejs.js": "./src/core/runtime/ipld-browser.js",
     "./test/utils/create-repo-nodejs.js": "./test/utils/create-repo-browser.js",
     "stream": "readable-stream",
     "joi": "joi-browser"
@@ -115,9 +116,11 @@
     "ipfs-unixfs-importer": "~0.38.5",
     "ipld": "~0.21.1",
     "ipld-bitcoin": "~0.1.8",
-    "ipld-dag-pb": "~0.15.3",
+    "ipld-dag-cbor": "^0.13.1",
+    "ipld-dag-pb": "^0.15.3",
     "ipld-ethereum": "^2.0.1",
     "ipld-git": "~0.3.0",
+    "ipld-raw": "^2.0.1",
     "ipld-zcash": "~0.1.6",
     "ipns": "~0.5.0",
     "is-ipfs": "~0.6.0",
